create ratelimter and upstash

+++ backend/server.js
@@ -80,20 +80,51 @@ app.delete("/api/transactions/:id", async (req, res) => {
     try {
         const { id } = req.params;
 
-        const result = await spl`
+        if(isNaN(parseInt(id))){
+            return res.status(400).json({ message: "Invalid transaction ID"});
+        }
+        const result = await sql`
         DELETE FROM transactions WHERE id = ${id} RETURNING *
         `;
-        if (result.length === 0 ){
+        if (result.length === 0){
             return res.status(404).json({ message: "Transaction not found"});
         }
 
-        res,status(200).json({ message: "Transaction deleted successfully"});
+        res.status(200).json({ message: "Transaction deleted successfully"});
     } catch (error) {
         console.log("Error deleting transaction", error);
         res.status(500).json({ message: "Internal Server Error"});
     }
 
 });
+
+app.get("/api/transactions/summary/:userId", async (req, res) => {
+    
+    try {
+        const {userId}=req.params;
+
+        const balanceResult = await sql`
+        SELECT COALESCE(SUM(amount),0) as balance FROM transactions WHERE user_id = ${userId}
+        `
+        const incomeResult = await sql`
+        SELECT COALESCE(SUM(amount),0) as income FROM transactions WHERE user_id = ${userId} AND amount > 0
+        `
+
+        const expenseResult = await sql`
+        SELECT COALESCE(SUM(amount),0) as expenses FROM transactions WHERE user_id = ${userId} AND amount < 0
+        `
+
+        res.status(200).json({
+            balance: balanceResult[0].balance,
+            income: incomeResult[0].income,
+            expenses: expenseResult[0].expenses
+        });
+    } catch (error) {
+        console.log("Error getting the summary", error);
+        res.status(500).json({ message: "Internal Server Error"});
+    }
+});
+
 
 initDB().then(() => {
     app.listen(PORT, () => {
